// src/app/generated/model/postReviewDto.ts
/**
 * swagger
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FilledReviewModuleDto } from './filledReviewModuleDto';


export interface PostReviewDto { 
    reviewId?: number;
    reviewModules?: Array<FilledReviewModuleDto> | null;
    grade?: string | null;
}