// src/app/generated/model/reviewDataDto.ts
/**
 * swagger
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ReviewModuleDto } from './reviewModuleDto';


export interface ReviewDataDto { 
    reviewId?: number;
    thesisId?: number;
    realizer?: string | null;
    supervisor?: string | null;
    topic?: string | null;
    modules?: Array<ReviewModuleDto> | null;
}